refactor: Auth 도메인 폴더 생성
