+++ client/src/pages/Auth/MyPage/components/SectionCard/SectionCard.jsx
