+++ client/src/pages/Auth/MyPage/components/SectionCard/SectionCard.module.css
