+++ client/src/pages/Auth/MyPage/components/TabContent/TabMyView/TabMyView.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState, useMemo } from 'react';
 import styles from './TabMyView.module.css';
-import { instance } from '../../../../../apis/instance';
+import { instance } from '../../../../../../apis/instance';
 import SectionCard from '../../SectionCard/SectionCard';
-import DropdownContainer from '../../../../../components/List/DropdownContainer/DropdownContainer';
-import tabMyViewFilter from '../../../../../utils/filters/tabMyViewFilter';
+import DropdownContainer from '../../../../../../components/List/DropdownContainer/DropdownContainer';
+import tabMyViewFilter from '../../../../../../utils/filters/tabMyViewFilter';
 import { useNavigate } from 'react-router-dom';
 import AttendanceModal from '../../AttendanceModal/AttendanceModal';
