--- client/src/pages/Test/Test.jsx
@@ -1,12 +1,0 @@
-import styles from './Test.module.css';
-import SearchList from './components/SearchList/SearchList';
-
-export default function Test() {
-  return (
-    <div className={styles.layout}>
-      <h1>아뜰리</h1>
-      테스트페이지입니다.
-      <SearchList />
-    </div>
-  );
-}
