--- client/src/pages/Test/Test.module.css
@@ -1,9 +1,0 @@
-.layout {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  min-height: calc(100vh - 4rem);
-
-  font-size: 1.6rem;
-}
