--- client/src/pages/Test/utils/filter.js
@@ -1,6 +1,0 @@
-export const filterList = [
-  { value: 'newest', label: '최신순' },
-  { value: 'oldest', label: '오래된순' },
-  { value: 'option3', label: 'Option3' },
-  { value: 'option4', label: 'Option4' },
-];
